+++ src/modules/products/infra/typeorm/entities/Product.ts
@@ -17,7 +17,7 @@ class Product {
   @Column()
   name: string;
 
-  @Column('decimal', { precision: 5, scale: 2 })
+  @Column('float', { precision: 5, scale: 2 })
   price: number;
 
   @Column()
